refactor: consolidate machine models
